Extract parser into its own phase
