+++ compiler/src/dotty/tools/dotc/interactive/InteractiveCompiler.scala
@@ -5,6 +5,7 @@ package interactive
 import core._
 import Phases._
 import typer._
+import parsing.Parsing
 
 class InteractiveCompiler extends Compiler {
   // TODO: Figure out what phases should be run in IDEs
@@ -12,6 +13,7 @@ class InteractiveCompiler extends Compiler {
   // This could be improved by reporting errors back to the IDE
   // after each phase group instead of waiting for the pipeline to finish.
   override def phases: List[List[Phase]] = List(
+    List(new Parsing),
     List(new FrontEnd),
     List(new transform.SetRootTree),
     List(new transform.CookComments)
