+++ doc-tool/src/dotty/tools/dottydoc/DocCompiler.scala
@@ -11,6 +11,7 @@ import dotc.typer.FrontEnd
 
 import dotty.tools.dotc.fromtasty.{ReadTasty, TASTYRun}
 import dotty.tools.dotc.transform.CookComments
+import dotty.tools.dotc.parsing.Parsing
 
 /** Custom Compiler with phases for the documentation tool
  *
@@ -53,6 +54,7 @@ class DocCompiler extends Compiler {
   }
 
   override def phases: List[List[Phase]] = List(
+    List(new Parsing),
     List(new DocFrontEnd),
     List(new ReadTasty),
     List(new CookComments),
